controller and routes change to single file
