--- backend/controller/patientController.js
@@ -1,69 +1,0 @@
-const Patient = require("../../model/receptionist/patient");
-
-// Register Patient
-exports.createPatient = async (req, res) => {
-  try {
-    const patient = new Patient(req.body);
-    await patient.save();
-    res.status(201).json(patient);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-};
-
-// Update Patient Info
-exports.updatePatient = async (req, res) => {
-  try {
-    const patient = await Patient.findOneAndUpdate(
-      { patientId: req.params.patientId },
-      req.body,
-      { new: true }
-    );
-    if (!patient) {
-      return res.status(404).json({ error: "Patient not found" });
-    }
-    res.json(patient);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-};
-
-// Get Patient by ID
-exports.getPatientById = async (req, res) => {
-  try {
-    const patient = await Patient.findOne({ patientId: req.params.patientId });
-    if (!patient) {
-      return res.status(404).json({ error: "Patient not found" });
-    }
-    res.json(patient);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-};
-
-// List All Patients
-exports.listPatients = async (req, res) => {
-  try {
-    const patients = await Patient.find();
-    res.json(patients);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-};
-
-// Deactivate Patient
-exports.deactivatePatient = async (req, res) => {
-  try {
-    const patient = await Patient.findOneAndUpdate(
-      { patientId: req.params.patientId },
-      { isActive: false },
-      { new: true }
-    );
-    if (!patient) {
-      return res.status(404).json({ error: "Patient not found" });
-    }
-    res.json(patient);
-  } catch (err) {
-    res.status(500).json({ error: err.message });
-  }
-};
