+++ backend/server.js
@@ -4,18 +4,14 @@ const mongoose = require("mongoose");
 require("dotenv").config({ path: __dirname + "/.env" });
 const cors = require("cors");
 
-const patientRoutes = require("./route/receptionist/patientRoutes");
-const appointmentRoutes = require("./route/receptionist/appointmentRoutes");
-const billingRoutes = require("./route/receptionist/billingRoutes");
+const receptionistRoutes = require("./route/receptionistRoutes");
 
 // Middleware
 app.use(cors());
 app.use(express.json());
 
 // Routes
-app.use("/api/patients", patientRoutes);
-app.use("/api/appointments", appointmentRoutes);
-app.use("/api/billing", billingRoutes);
+app.use("/api/receptionist", receptionistRoutes);
 
 // Start server
 const PORT = process.env.PORT || 8000;
